Fix UB which led to an infinite loop due to using `nconc` on the result of a backquote
Prior to this commit, recursive calls to `collect-maybe-itemized-section` due to nested
lists would call `nconc` to concatenate the result of the recursive call. That result was
produced by a backquote template whose final item was not a splicing-unquote. As a result,
the `nconc` would mutate the template, so future calls to `collect-maybe-itemized-section`
would either loop infinitely or return a circular list.